server ingredient score calculation working
